+++ CHANGELOG.md
@@ -33,3 +33,4 @@
 - apply spotless
 ## 27-03-2026-1.4.5 Modified by xXjudeXx
 - update version number to 1.4.5
+- Add Redstone, Glowstone & Ender Metals